Moved assignment tests to new file.
